src/test/java/com/leetcode/strings/easy/RotateString_796Test.java
package com.leetcode.strings.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class RotateString_796Test {

  private final RotateString_796 solution = new RotateString_796();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments("abcde", "cdeab", true), arguments("abcde", "abced", false));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void rotateString(String str1, String str2, boolean expectedResult) {
    boolean actualResult = solution.rotateString(str1, str2);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
